+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `most_variable_features` method
 - `characteristic_features` method
 - `most_correlated` method
+- MissingFilter structures 
 
 ## v1.0.x - unreleased
 ### Changed